# src/config/swagger.yaml
swagger: "2.0"
info:
  description: "Controle Simples de Postagens com Usuários"
  version: "1.0.0"
  title: "posts_control"
basePath: "/api/v1"
consumes:
  - "application/json"
produces:
  - "application/json"
...
definitions:
  user:
    additionalProperties: false
    type: "object"
    required:
      - "full_name"
      - "user_email"
      - "user_password"
    properties:
      id:
        type: "integer"
      full_name:
        type: "string"
      user_email:
        type: "string"
      user_password:
        type: "string"
...
paths:
...
  /user/{user_id}:
    put:
      tags:
        - "Users"
      x-swagger-router-controller: app
      summary: update user
      operationId: updateUserHandler
      parameters:
        - name: user_id
          in: path
          type: "integer"
          required: true
        - in: body
          name: user
          description: update user
          schema:
            additionalProperties: false
            type: "object"
            required:
              - "full_name"
              - "user_email"
              - "user_password"
            properties:
              full_name:
                type: "string"
              user_email:
                type: "string"
              user_password:
                type: "string"
      responses:
        200:
          description: "success"
          schema:
            $ref: "#/definitions/user"
    delete:
      tags:
        - "Users"
      x-swagger-router-controller: app
      summary: delete user
      operationId: deleteUserHandler
      parameters:
        - name: user_id
          in: path
          type: integer
          required: true
      responses:
        200:
          description: "success"
          schema:
            $ref: "#/definitions/user"
...